Migrated Twitter lib function(s) to TypeScript
